more changes to finish

+++ app/src/main/java/edu/orangecoastcollege/cs273/wheretonext/CollegeDetailsActivity.java
@@ -32,8 +32,8 @@ public class CollegeDetailsActivity extends AppCompatActivity {
         Intent detailsIntent = getIntent();
         String name = detailsIntent.getStringExtra("Name");
         int population = detailsIntent.getIntExtra("Population", 0);
-        float tuition = detailsIntent.getFloatExtra("Tuition", 0.0f);
-        float rating = detailsIntent.getFloatExtra("Rating", 0.0f);
+        double tuition = detailsIntent.getDoubleExtra("Tuition", 0.0f);
+        double rating = detailsIntent.getDoubleExtra("Rating", 0.0f);
         String imageName = detailsIntent.getStringExtra("ImageName");
 
         AssetManager am = this.getAssets();
@@ -53,7 +53,7 @@ public class CollegeDetailsActivity extends AppCompatActivity {
         collegeDetailsNameTextView.setText(name);
         collegeDetailsPopulationTextView.setText("Annual Enrollment: " + thousands.format(population));
         collegeDetailsTuitionTextView.setText("In-state Tuition: " + currency.format(tuition));
-        gameDetailsRatingBar.setRating(rating);
+        gameDetailsRatingBar.setRating((float)rating);
     }
